+++ fmsetlist.v
@@ -40,7 +40,7 @@ Section FMSet2List.
 
   Definition list2fm l := fold_right add empty l.
 
-  Global Instance list2fm_perm : Proper (@Permutation A ==> eq) list2fm
+  #[export] Instance list2fm_perm : Proper (@Permutation A ==> eq) list2fm
     := perm_eq.
 
   Lemma list2fm_retract m : list2fm (elts m) = m.
@@ -67,7 +67,7 @@ Section FMSet2List.
   - now rewrite elts_add, IHl.
   Qed.
 
-  Global Instance elts_perm' : Proper (eq ==> @Permutation A) elts.
+  #[export] Instance elts_perm' : Proper (eq ==> @Permutation A) elts.
   Proof. now intros m1 m2 Heq; subst. Qed.
 
   Lemma elts_eq_nil m : elts m = nil -> m = empty.